RAG/requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
faiss-cpu==1.8.0.post1
pydantic>=2.7.0
python-dotenv>=1.0.1
langchain-ollama>=0.1.0
